requirements_anomaly.txt
torch>=2.0.0
transformers>=4.36.0
numpy>=1.24.0
Pillow>=9.0.0
